+++ backend/src/main/java/mrs/eclinicapi/service/ClinicRoomService.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 public interface ClinicRoomService {
         List<ClinicRoom> getAllClinicRooms();
-	void remove(int id);
-	ClinicRoom findOneClinicRoom(int id);
+	void remove(Long id);
+	ClinicRoom findOneClinicRoom(Long id);
 }
